pyproject.toml
[project]
name = "violence-monitoring"
version = "0.1.0"
description = "Video violence monitoring with YOLO and ViT pipelines"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "timm>=1.0.26",
    "torch>=2.11.0",
    "transformers>=5.5.0",
    "ultralytics>=8.4.33",
]

[project.scripts]
run-yolo-monitor = "violence_detection.yolo_detector_monitor:main"
run-vit-monitor = "violence_detection.vit_classifier_monitor:main"
